Stability and security improved

+++ utils/dbDialog.js
@@ -2,7 +2,7 @@ const mysql = require("mysql");
 const config = require("../config.json");
 
 module.exports = dbDialog = {
-    connectDB: async function() {
+    connectDB: async function () {
         this.db = await mysql.createConnection({
             host: config.db.host,
             user: config.db.user,
@@ -10,13 +10,15 @@ module.exports = dbDialog = {
             database: config.db.db
         });
 
-        this.db.connect(function(err) {
-            if(err) throw err;
+        this.db.connect(function (err) {
+            if (err) throw err;
         });
     },
-    insertUser: async function(discordID, creepName) {
+    insertUser: async function (userId, creeptdUsername) {
+        if (!(typeof (userId) == 'string' && typeof (creeptdUsername) == 'string')) throw new Error('Variable type error');
+
         const sql = "INSERT INTO users(discord_id, creeptd_username) VALUES(?, ?)";
-        let results = await new Promise((resolve, reject) => this.db.query(sql, [discordID, creepName], (err, results) => {
+        let results = await new Promise((resolve, reject) => this.db.query(sql, [userId, creeptdUsername], (err, results) => {
             if (err) {
                 reject(err);
             }
@@ -25,9 +27,11 @@ module.exports = dbDialog = {
         }));
         return results;
     },
-    deleteUser: async function(discordID) {
+    deleteUser: async function (userId) {
+        if (!(typeof (userId) == 'string')) throw new Error('Variable type error');
+
         const sql = "DELETE FROM users WHERE discord_id=?";
-        let results = await new Promise((resolve, reject) => this.db.query(sql, [discordID], (err, results) => {
+        let results = await new Promise((resolve, reject) => this.db.query(sql, [userId], (err, results) => {
             if (err)
                 reject(err);
             else
@@ -35,25 +39,25 @@ module.exports = dbDialog = {
         }));
         return results;
     },
-    getCreepName: async function(discordID) {
+    getCreepName: async function (userId) {
+        if (!(typeof (userId) == 'string')) throw new Error('Variable type error');
+
         const sql = "SELECT creeptd_username FROM users WHERE discord_id=?";
-        let results = await new Promise((resolve, reject) => this.db.query(sql, [discordID], (err, results) => {
+        let results = await new Promise((resolve, reject) => this.db.query(sql, [userId], (err, results) => {
             if (err)
                 reject(err);
-            else {
-                if (results.length < 1)
-                    reject(new Error('NO_DATA'));
-                else
-                    resolve(results);
-            }
+            else if (results.length < 1)
+                reject(new Error('NO_DATA'));
+            else
+                resolve(results);
         }));
-        if (results && results[0] && results[0].creeptd_username)
-            return results[0].creeptd_username;
-        return results;
+        return results[0].creeptd_username;
     },
-    insertGuild: async function(guildID) {
+    insertGuild: async function (guildId) {
+        if (!(typeof (guildId) == 'string')) throw new Error('Variable type error');
+
         let sql = "INSERT INTO guilds(discord_id, language) VALUES(?, 0)";
-        let result = await new Promise((resolve, reject) => this.db.query(sql, [guildID], (err, results) => {
+        let result = await new Promise((resolve, reject) => this.db.query(sql, [guildId], (err, results) => {
             if (err)
                 reject(err);
             else
@@ -61,35 +65,25 @@ module.exports = dbDialog = {
         }));
         return result;
     },
-    updateLanguage: async function(guildID, languageID) {
+    updateLanguage: async function (guildId, languageCode) {
+        if (!(typeof (guildId) == 'string' && typeof (languageCode) == 'string')) throw new Error('Variable type error');
+
         let sql = "UPDATE guilds SET language=(SELECT id FROM languages WHERE code=?) WHERE discord_id=?";
-        let results = await new Promise((resolve, reject) => this.db.query(sql, [languageID, guildID], async (err, results) => {
+        let results = await new Promise((resolve, reject) => this.db.query(sql, [languageCode, guildId], async (err, results) => {
             if (err)
                 reject(err);
-            else {
-                if (results.affectedRows === 0) {
-                    sql = "INSERT INTO guilds(discord_id, language) VALUES(?, (SELECT id FROM languages WHERE code=?))";
-                    results = await new Promise((resolve, reject) => this.db.query(sql, [guildID, languageID], (err, results) => {
-                        if (err) {
-                            reject(err);
-                        }
-                        else {
-                            resolve(results);
-                        }
-                    }));
-                }
-                else if (results.changedRows === 0) {
-                    reject(new Error('NO_DATA_CHANGED'));
-                }
-                else
-                    resolve(results);
-            }
+            if (results.changedRows === 0)
+                reject(new Error('NO_DATA_CHANGED'));
+            else
+                resolve(results);
         }));
         return results;
     },
-    getLanguage: async function(guildID) {
+    getLanguage: async function (guildId) {
+        if (!(typeof (guildId) == 'string')) throw new Error('Variable type error');
+
         let sql = "SELECT LA.code FROM languages LA JOIN guilds GU ON LA.id=GU.language WHERE GU.discord_id=?";
-        let results = await new Promise((resolve, reject) => this.db.query(sql, [guildID], (err, results) => {
+        let results = await new Promise((resolve, reject) => this.db.query(sql, [guildId], (err, results) => {
             if (err)
                 reject(err);
             else {
@@ -99,8 +93,6 @@ module.exports = dbDialog = {
                     resolve(results);
             }
         }));
-        if (results && results[0] && results[0].code)
-            return results[0].code;
-        return results;
+        return results[0].code;
     }
 }
